requirements.txt
# Core ML and NLP libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
nltk>=3.6.0
spacy>=3.0.0

# Deep Learning (optional, for advanced models)
torch>=1.9.0
transformers>=4.11.0

# Text processing
textblob>=0.15.3
emoji>=1.7.0

# Data visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
pyyaml>=5.4.0
tqdm>=4.62.0
joblib>=1.0.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
